When disconnected, userinfo thread stops fetching userinfo (clutters log otherwise).
When an exchange is opened, disable reconnecting.
This fixes a bug where if the internet or host is not working, it will
attempt to reconnect and fail, then the on_close will set reconnect to
true, but when the internet is re-enabled, exchange_handler will
immediately reconnect because reconnect was never set to false.